Make clear in README that this is a django reusable-app (#67)
Make clear in README that this is a django reusable-app, not a standalone application, or a framework, or a library.

The README states that this is an app, but to someone not familiar with django idioms the term is easily misleading. Using a ubiquitous language to talk about what this project is in relation to the Python ecosystem and link to documentation about what type of thing this is will help prevent some misunderstandings.

Specifically this will help preventing someone like myself making the mistake I made.